refactor(error):errorName 상수화, 파일명변경
